adding registration functionality
connecting to register api and adapting form to interact with user. the form is responsive and gives user feedback. once valid parameters are submitted the response is checked. if successful user is notified of the successful registration and redirected to login page. if not, the user is notified of the error